requirements.txt
Django==2.2.13
django-cors-headers==3.0.2
djangorestframework==3.10.2
jsonschema==3.0.1
lxml==4.3.4
Markdown==3.1.1
psycopg2==2.8.3
pyOpenSSL==19.0.0
requests==2.22.0
sqlparse==0.3.0
virtualenv==16.6.2
Werkzeug==0.15.4
oauth2client==4.1.3